+++ jannovar-cli/src/main/java/de/charite/compbio/jannovar/cmd/annotate_vcf/AnnotateVCFCommand.java
@@ -105,19 +105,22 @@ public class AnnotateVCFCommand extends JannovarAnnotationCommand {
 		final String vcfPath = options.getPathInputVCF();
 
 		// whether or not to require availability of an index
-		final boolean useInterval = (options.getInterval() != null && !options.getInterval().equals(""));
+		final boolean useInterval = (options.getInterval() != null
+				&& !options.getInterval().equals(""));
 
 		try (VCFFileReader vcfReader = new VCFFileReader(new File(vcfPath), useInterval)) {
 			if (this.options.getVerbosity() >= 1) {
-				final SAMSequenceDictionary seqDict = VCFFileReader.getSequenceDictionary(new File(vcfPath));
+				final SAMSequenceDictionary seqDict = VCFFileReader
+						.getSequenceDictionary(new File(vcfPath));
 				if (seqDict != null) {
 					final GenomeRegionListFactoryFromSAMSequenceDictionary factory = new GenomeRegionListFactoryFromSAMSequenceDictionary();
 					this.progressReporter = new ProgressReporter(factory.construct(seqDict), 60);
 					this.progressReporter.printHeader();
 					this.progressReporter.start();
 				} else {
-					System.err.println("Progress reporting does not work because VCF file is missing the contig "
-							+ "lines in the header.");
+					System.err.println(
+							"Progress reporting does not work because VCF file is missing the contig "
+									+ "lines in the header.");
 				}
 			}
 
@@ -176,7 +179,8 @@ public class AnnotateVCFCommand extends JannovarAnnotationCommand {
 				DBAnnotationOptions gnomadOptions = DBAnnotationOptions.createDefaults();
 				gnomadOptions.setIdentifierPrefix(options.prefixGnomadExomes);
 				DBVariantContextAnnotator gnomadExomesAnno = new DBVariantContextAnnotatorFactory()
-						.constructGnomad(options.pathVCFGnomadExomes, options.pathFASTARef, gnomadOptions);
+						.constructGnomad(options.pathVCFGnomadExomes, options.pathFASTARef,
+								gnomadOptions);
 				gnomadExomesAnno.extendHeader(vcfHeader);
 				stream = stream.map(gnomadExomesAnno::annotateVariantContext);
 			}
@@ -188,9 +192,22 @@ public class AnnotateVCFCommand extends JannovarAnnotationCommand {
 				DBAnnotationOptions gnomadOptions = DBAnnotationOptions.createDefaults();
 				gnomadOptions.setIdentifierPrefix(options.prefixGnomadGenomes);
 				DBVariantContextAnnotator gnomadGenomesAnno = new DBVariantContextAnnotatorFactory()
-						.constructGnomad(options.pathVCFGnomadGenomes, options.pathFASTARef, gnomadOptions);
+						.constructGnomad(options.pathVCFGnomadGenomes, options.pathFASTARef,
+								gnomadOptions);
 				gnomadGenomesAnno.extendHeader(vcfHeader);
 				stream = stream.map(gnomadGenomesAnno::annotateVariantContext);
+			}
+
+			// If configured, annotate using thousand genomes VCF file (extend
+			// header to use for writing out)
+			if (options.pathThousandGenomes != null) {
+				DBAnnotationOptions thousandGenomesOptions = DBAnnotationOptions.createDefaults();
+				thousandGenomesOptions.setIdentifierPrefix(options.prefixThousandGenomes);
+				DBVariantContextAnnotator thousandGenomesAnno = new DBVariantContextAnnotatorFactory()
+						.constructThousandGenomes(options.pathThousandGenomes, options.pathFASTARef,
+								thousandGenomesOptions);
+				thousandGenomesAnno.extendHeader(vcfHeader);
+				stream = stream.map(thousandGenomesAnno::annotateVariantContext);
 			}
 
 			// If configured, annotate using UK10K VCF file (extend header to
@@ -210,7 +227,8 @@ public class AnnotateVCFCommand extends JannovarAnnotationCommand {
 				DBAnnotationOptions clinVarOptions = DBAnnotationOptions.createDefaults();
 				clinVarOptions.setIdentifierPrefix(options.prefixClinVar);
 				DBVariantContextAnnotator clinvarAnno = new DBVariantContextAnnotatorFactory()
-						.constructClinVar(options.pathClinVar, options.pathFASTARef, clinVarOptions);
+						.constructClinVar(options.pathClinVar, options.pathFASTARef,
+								clinVarOptions);
 				clinvarAnno.extendHeader(vcfHeader);
 				stream = stream.map(clinvarAnno::annotateVariantContext);
 			}
@@ -229,14 +247,15 @@ public class AnnotateVCFCommand extends JannovarAnnotationCommand {
 			// Add step for annotating with variant effect
 			VariantEffectHeaderExtender extender = new VariantEffectHeaderExtender();
 			extender.addHeaders(vcfHeader);
-			VariantContextAnnotator variantEffectAnnotator =
-					new VariantContextAnnotator(refDict, chromosomeMap,
-							new VariantContextAnnotator.Options(!options.isShowAll(), 
-									(options.isUseThreeLetterAminoAcidCode() ? AminoAcidCode.THREE_LETTER : AminoAcidCode.ONE_LETTER),
-									options.isEscapeAnnField(), options.isNt3PrimeShifting(),
-									options.isOffTargetFilterEnabled(),
-									options.isOffTargetFilterUtrIsOffTarget(),
-									options.isOffTargetFilterIntronicSpliceIsOffTarget()));
+			VariantContextAnnotator variantEffectAnnotator = new VariantContextAnnotator(refDict,
+					chromosomeMap,
+					new VariantContextAnnotator.Options(!options.isShowAll(),
+							(options.isUseThreeLetterAminoAcidCode() ? AminoAcidCode.THREE_LETTER
+									: AminoAcidCode.ONE_LETTER),
+							options.isEscapeAnnField(), options.isNt3PrimeShifting(),
+							options.isOffTargetFilterEnabled(),
+							options.isOffTargetFilterUtrIsOffTarget(),
+							options.isOffTargetFilterIntronicSpliceIsOffTarget()));
 			stream = stream.map(variantEffectAnnotator::annotateVariantContext);
 
 			// If configured, use threshold-based annotation (extend header to
@@ -248,10 +267,15 @@ public class AnnotateVCFCommand extends JannovarAnnotationCommand {
 						options.getThreshFiltMinGtCovHet(), options.getThreshFiltMinGtCovHomAlt(),
 						options.getThreshFiltMaxCov(), options.getThreshFiltMinGtGq(),
 						options.getThreshFiltMinGtAafHet(), options.getThreshFiltMaxGtAafHet(),
-						options.getThreshFiltMinGtAafHomAlt(), options.getThreshFiltMaxGtAafHomRef(),
-						options.getPrefixExac(), options.getPrefixDBSNP(), options.getPrefixGnomadGenomes(),
-						options.getPrefixGnomadExomes(), options.getThreshFiltMaxAlleleFrequencyAd(),
-						options.getThreshFiltMaxAlleleFrequencyAr());
+						options.getThreshFiltMinGtAafHomAlt(),
+						options.getThreshFiltMaxGtAafHomRef(), options.getPrefixExac(),
+						options.getPrefixDBSNP(), options.getPrefixGnomadGenomes(),
+						options.getPrefixGnomadExomes(),
+						options.getPrefixThousandGenomes(),
+						options.getThreshFiltMaxAlleleFrequencyAd(),
+						options.getThreshFiltMaxAlleleFrequencyAr(),
+						options.getThreshFiltMaxExacHomAlt(),
+						options.getThreshFiltMaxThousandGenomesHomAlt());
 				// Add headers
 				new ThresholdFilterHeaderExtender(thresholdFilterOptions).addHeaders(vcfHeader);
 				// Build list of affecteds; take from pedigree file if given.
@@ -286,8 +310,8 @@ public class AnnotateVCFCommand extends JannovarAnnotationCommand {
 										+ "only genotype FT");
 					}
 				}
-				GenotypeThresholdFilterAnnotator gtThresholdFilterAnno =
-						new GenotypeThresholdFilterAnnotator(thresholdFilterOptions);
+				GenotypeThresholdFilterAnnotator gtThresholdFilterAnno = new GenotypeThresholdFilterAnnotator(
+						thresholdFilterOptions);
 				stream = stream.map(gtThresholdFilterAnno::annotateVariantContext);
 
 				// When configured to use advanced pedigree filters (must come
@@ -312,14 +336,14 @@ public class AnnotateVCFCommand extends JannovarAnnotationCommand {
 					}
 
 					// Construct annotator and register with pipeline
-					PedigreeFilterAnnotator pedFilterAnnotator = new PedigreeFilterAnnotator(pedFilterOptions,
-							pedigree);
+					PedigreeFilterAnnotator pedFilterAnnotator = new PedigreeFilterAnnotator(
+							pedFilterOptions, pedigree);
 					stream = stream.map(pedFilterAnnotator::annotateVariantContext);
 				}
 
 				if (options.useThresholdFilters) {
-					VariantThresholdFilterAnnotator varThresholdFilterAnno =
-							new VariantThresholdFilterAnnotator(thresholdFilterOptions, affecteds);
+					VariantThresholdFilterAnnotator varThresholdFilterAnno = new VariantThresholdFilterAnnotator(
+							thresholdFilterOptions, affecteds);
 					stream = stream.map(varThresholdFilterAnno::annotateVariantContext);
 				}
 			}
@@ -340,21 +364,24 @@ public class AnnotateVCFCommand extends JannovarAnnotationCommand {
 				for (String colName : options.getColumnsDbNsfp()) {
 					descriptions.put(colName, DbNsfpFields.DBNSFP_FIELDS.get(colName));
 				}
-				GenericTSVAnnotationOptions dbNsfpAnnotationOptions = new GenericTSVAnnotationOptions(true, false,
-						options.getPrefixDbNsfp(), MultipleMatchBehaviour.BEST_ONLY, new File(options.getPathDbNsfp()),
-						GenericTSVAnnotationTarget.VARIANT, true, options.getDbNsfpColContig(),
-						options.getDbNsfpColPosition(), options.getDbNsfpColPosition(), 3, 4, false, 
-						options.getColumnsDbNsfp(), descriptions);
-				dbNsfpAnnotator = new GenericTSVAnnotationDriver(options.getPathFASTARef(), dbNsfpAnnotationOptions);
+				GenericTSVAnnotationOptions dbNsfpAnnotationOptions = new GenericTSVAnnotationOptions(
+						true, false, options.getPrefixDbNsfp(), MultipleMatchBehaviour.BEST_ONLY,
+						new File(options.getPathDbNsfp()), GenericTSVAnnotationTarget.VARIANT, true,
+						options.getDbNsfpColContig(), options.getDbNsfpColPosition(),
+						options.getDbNsfpColPosition(), 3, 4, false, options.getColumnsDbNsfp(),
+						descriptions);
+				dbNsfpAnnotator = new GenericTSVAnnotationDriver(options.getPathFASTARef(),
+						dbNsfpAnnotationOptions);
 				dbNsfpAnnotator.constructVCFHeaderExtender().addHeaders(vcfHeader);
 				stream = stream.map(dbNsfpAnnotator::annotateVariantContext);
 			}
 
 			// Annotate from generic TSV files
 			List<GenericTSVAnnotationDriver> tsvAnnotators = new ArrayList<>();
-			for (GenericTSVAnnotationOptions tsvAnnotationOptions : options.getTsvAnnotationOptions()) {
-				GenericTSVAnnotationDriver annotator = new GenericTSVAnnotationDriver(options.getPathFASTARef(),
-						tsvAnnotationOptions);
+			for (GenericTSVAnnotationOptions tsvAnnotationOptions : options
+					.getTsvAnnotationOptions()) {
+				GenericTSVAnnotationDriver annotator = new GenericTSVAnnotationDriver(
+						options.getPathFASTARef(), tsvAnnotationOptions);
 				tsvAnnotators.add(annotator);
 				annotator.constructVCFHeaderExtender().addHeaders(vcfHeader);
 				stream = stream.map(annotator::annotateVariantContext);
@@ -362,9 +389,11 @@ public class AnnotateVCFCommand extends JannovarAnnotationCommand {
 
 			// Annotate from generic VCF files
 			List<GenericVCFAnnotationDriver> vcfAnnotators = new ArrayList<>();
-			for (GenericVCFAnnotationOptions vcfAnnotationOptions : options.getVcfAnnotationOptions()) {
+			for (GenericVCFAnnotationOptions vcfAnnotationOptions : options
+					.getVcfAnnotationOptions()) {
 				GenericVCFAnnotationDriver annotator = new GenericVCFAnnotationDriver(
-						vcfAnnotationOptions.getPathVcfFile(), options.getPathFASTARef(), vcfAnnotationOptions);
+						vcfAnnotationOptions.getPathVcfFile(), options.getPathFASTARef(),
+						vcfAnnotationOptions);
 				vcfAnnotators.add(annotator);
 				annotator.constructVCFHeaderExtender().addHeaders(vcfHeader);
 				stream = stream.map(annotator::annotateVariantContext);
@@ -400,8 +429,8 @@ public class AnnotateVCFCommand extends JannovarAnnotationCommand {
 					(endTime - startTime) / 1000.0 / 1000.0 / 1000.0));
 		} catch (IncompatiblePedigreeException e) {
 			if (options.pathPedFile != null)
-				System.err
-						.println("VCF file " + vcfPath + " is not compatible to pedigree file " + options.pathPedFile);
+				System.err.println("VCF file " + vcfPath + " is not compatible to pedigree file "
+						+ options.pathPedFile);
 			else
 				System.err.println("VCF file " + vcfPath
 						+ " is not compatible with singleton pedigree annotation (do you have exactly one sample in VCF file?)");
@@ -425,9 +454,11 @@ public class AnnotateVCFCommand extends JannovarAnnotationCommand {
 	/**
 	 * Load pedigree from file given in configuration or construct singleton pedigree
 	 * 
-	 * @param vcfHeader {@link VCFHeader}, for checking compatibility and getting sample name in
-	 *        case of singleton pedigree construction
-	 * @throws PedParseException in the case of problems with parsing pedigrees
+	 * @param vcfHeader
+	 *            {@link VCFHeader}, for checking compatibility and getting sample name in case of
+	 *            singleton pedigree construction
+	 * @throws PedParseException
+	 *             in the case of problems with parsing pedigrees
 	 */
 	private Pedigree loadPedigree(VCFHeader vcfHeader)
 			throws PedParseException, IOException, IncompatiblePedigreeException {
@@ -440,10 +471,10 @@ public class AnnotateVCFCommand extends JannovarAnnotationCommand {
 				throw new IncompatiblePedigreeException(
 						"VCF file does not have exactly one sample but required for singleton pedigree construction");
 			final String sampleName = vcfHeader.getSampleNamesInOrder().get(0);
-			final PedPerson pedPerson =
-					new PedPerson(sampleName, sampleName, "0", "0", Sex.UNKNOWN, Disease.AFFECTED);
-			final PedFileContents pedContents =
-					new PedFileContents(ImmutableList.of(), ImmutableList.of(pedPerson));
+			final PedPerson pedPerson = new PedPerson(sampleName, sampleName, "0", "0", Sex.UNKNOWN,
+					Disease.AFFECTED);
+			final PedFileContents pedContents = new PedFileContents(ImmutableList.of(),
+					ImmutableList.of(pedPerson));
 			return new Pedigree(pedContents, pedContents.getIndividuals().get(0).getPedigree());
 		}
 	}
@@ -451,12 +482,17 @@ public class AnnotateVCFCommand extends JannovarAnnotationCommand {
 	/**
 	 * Construct the mendelian inheritance annotation processors
 	 * 
-	 * @param writer the place to put put the VariantContext to after filtration
-	 * @param vcfHeader {@link VCFHeader}, for checking compatibility and getting sample name in
-	 *        case of singleton pedigree construction
-	 * @throws IOException in case of problems with opening the pedigree file
-	 * @throws PedParseException in the case of problems with parsing pedigrees
-	 * @throws IncompatiblePedigreeException If the pedigree is incompatible with the VCF file
+	 * @param writer
+	 *            the place to put put the VariantContext to after filtration
+	 * @param vcfHeader
+	 *            {@link VCFHeader}, for checking compatibility and getting sample name in case of
+	 *            singleton pedigree construction
+	 * @throws IOException
+	 *             in case of problems with opening the pedigree file
+	 * @throws PedParseException
+	 *             in the case of problems with parsing pedigrees
+	 * @throws IncompatiblePedigreeException
+	 *             If the pedigree is incompatible with the VCF file
 	 */
 	private VariantContextProcessor buildMendelianProcessors(VariantContextWriter writer,
 			VCFHeader vcfHeader)
@@ -464,9 +500,9 @@ public class AnnotateVCFCommand extends JannovarAnnotationCommand {
 		if (options.pathPedFile != null || options.annotateAsSingletonPedigree) {
 			final Pedigree pedigree = loadPedigree(vcfHeader);
 			checkPedigreeCompatibility(pedigree, vcfHeader);
-			final GeneWiseMendelianAnnotationProcessor mendelProcessor =
-					new GeneWiseMendelianAnnotationProcessor(pedigree, jannovarData,
-							vc -> writer.add(vc), options.isInheritanceAnnoUseFilters());
+			final GeneWiseMendelianAnnotationProcessor mendelProcessor = new GeneWiseMendelianAnnotationProcessor(
+					pedigree, jannovarData, vc -> writer.add(vc),
+					options.isInheritanceAnnoUseFilters());
 			return new CoordinateSortingChecker(mendelProcessor);
 		} else {
 			return new ConsumerProcessor(vc -> writer.add(vc));
@@ -476,17 +512,21 @@ public class AnnotateVCFCommand extends JannovarAnnotationCommand {
 	/**
 	 * Check pedigree for compatibility
 	 * 
-	 * @param pedigree {@link Pedigree} to check for compatibility
-	 * @param vcfHeader {@link VCFHeader} to check for compatibility
-	 * @throws IncompatiblePedigreeException if the VCF file is not compatible with the pedigree
+	 * @param pedigree
+	 *            {@link Pedigree} to check for compatibility
+	 * @param vcfHeader
+	 *            {@link VCFHeader} to check for compatibility
+	 * @throws IncompatiblePedigreeException
+	 *             if the VCF file is not compatible with the pedigree
 	 */
 	private void checkPedigreeCompatibility(Pedigree pedigree, VCFHeader vcfHeader)
 			throws IncompatiblePedigreeException {
 		List<String> missing = vcfHeader.getGenotypeSamples().stream()
 				.filter(x -> !pedigree.getNames().contains(x)).collect(Collectors.toList());
-		if (!missing.isEmpty()) throw new IncompatiblePedigreeException(
-				"The VCF file has the following sample names not present in Pedigree: "
-						+ Joiner.on(", ").join(missing));
+		if (!missing.isEmpty())
+			throw new IncompatiblePedigreeException(
+					"The VCF file has the following sample names not present in Pedigree: "
+							+ Joiner.on(", ").join(missing));
 	}
 
 }